pyproject.toml
[project]
name = "astroviewer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Momo", email = "39585852+imoken1122@users.noreply.github.com" }
]
dependencies = [
    "flet>=0.10.1",
    "numpy>=1.25.2",
    "pillow>=10.0.1",
    "requests>=2.31.0",
    "pysvb>=0.2.0",
    "astropy>=5.3.3",
    "uvicorn>=0.23.2",
    "fastapi>=0.103.1",
    "starlette<0.28.0,>=0.27.0",
    "paho-mqtt>=1.6.1",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
